chore!: move to util
